Remove error log when gas < intrinsic gas.
This log statement was being hit as part of gas estimation if using non-Celo currencies for the fees.  However, this is a normal part of gas estimation's binary search, which sets the search's bottom at 21k gas and doesn't account for intrinsic gas being higher in some cases.  As such, it should not log an error message.  An alternative solution would be to calculate the intrinsic gas taking into account the fee currency and use that as the low end of the binary search.