Decouples EVM from the contract layer (#1465)
* EVM: Removes knowledge of TobinTax

EVM should be unaware of tobin tax. From the EVM perspective there is
just the Context.Transfer() function.

This commit, moves evm.TobinTransfer to replace context.Transfer
function. In so, it removes unnecessary input and output parameters.

* EVM: Removes static call to GetRegisteredAddress

This commit removes the use of a static call to obtain a registry
address. This was used on the Transfer precompile to check that the only
caller to the contract is the GoldToken contract.

With this commit, there's a GetRegisteredAddress function on the
evm.Context that is injected at the moment of creating.

For now, this function is hardcoded, but later it will be received as a
parameter to the vm.NewContext()

* EVM module: Removes GetRegistedAddressWithEVM

Simple refactor, replaces calls to GetRegistedAddressWithEVM to the new
version GetRegistedAddress.

As a side effect, return type was change from *common.Address to
common.Address which might be solving a bug, since we found that we are
comparing two pointers when we should be comparing values. (on
currency.Convert())

* Removes notion of system call & abi on the evm

celo-blockchain added `StaticCallFromSystem` and `CallFromSystem` which
also implied knowledge of the abi module and of the system caller
(address zero)

Both concepts don't belong in the EVM and are better placed on a higher
layer that "uses" the EVM.

This commit extracts those methods, into the `contracts` module.
Decouples the evm from those concept and leaves it closer to upstream's
EVM.

Additionally, `GetRegisteredAddress` is moved to the contracts module.
Decoupling the evm module from the concept of "Registry" and of a
"Registry Contract".

This commit, introduces a new parameter to the vm.NewContext():
ContractsContext which is an interface that provides a
`GetRegisteredAddress` method for use within the EVM. It is a interface
instead of a single function since `getTobinTax` will also be included
there in a follow up PR.

* Moves vm.NewContext to context.New

With this commit, we break the circular dependency between contracts &
vm by moving out the creation of the Context object. This object needs
to define a function to get registered address and a function to
compute the tobin tax.

Alternatively (as in a previous commit) we can get that separation by
injection those function in the ContractsContext. But this approach is
simpler, and also completely removes the notion of Tobin tax from the vm
package.

* Extracts TobinTax computation to contracts pkg

Computing the tobin tax implies a contract call. As such, it is better
placed on contract pkg.

* Moves tobin tax to contracts/reserve pkg

Additionally, changes output to Ratio.

Fixes lint warning about using Capitalized strings on error text

* Rename context to vmcontext

* nit

* Nit changes to move closer to upstream

* Reverts old changes to go back to upstream

On upstream:

 * There's core.NewEVMContext
 * There's core.Message

I moved/recreated those 2 artifacts in their original location. The main
difference is that `core.NewEVMContext` will call `vmcontext.New()`
instead of creating the context itself.

`vm.ChainContext` was actually `core.ChainContext` but for the moment
i've left that in the current location.

* fix pr comment

* Makes  reserve.Ratio fields private

* Move evm closer to upstream.

Removes unnecessary methods

* Reverts change. fix test