Removes check on feeCurrency == goldTokenAddress
On master this was only ever used on ConvertToGold which was called on
tx_list. It was used as part of the logic to compare two txs with
different fee currencies, and this logic it's just a shortcut.

In reality, goldtokenAddress is not a whitelisted fee currency, since it's
expressed with feeCurrency = nil