Only active validator signs stats (#1508)
Description
A proxy with two validators (one active) that send a stat to be signed by the validator, was receiving 2 signed messages for the same (one per validator)

Only the active validator will sign those stats.

Other changes
The validators MUST have the celostats flag too (as we have it in the docs)

Tested
Manually

Backwards compatibility
Breaking change in that proxied validators need to specify the celostats flag in order to report to celostats. However, this new behavior agrees with our documentation, so this is really more of a bugfix.