Make --proxy.proxyenodeurlpairs optional if --proxy.proxied specified (#1498)
### Description

Documentation implies this behavior is intended. For example,
https://docs.celo.org/validator-guide/proxy.

This facilitates running geth without initially configuring a proxy and adding
proxies later via RPCs. For example, like
https://github.com/mantalabs/proxy-informer.

### Tested

Ad hoc testing locally.

### Backwards compatibility

This change is technically a breaking one because it changes the behavior of the
(deprecated) --proxy.proxyenodeurlpair option from required to
optional. This will break uses where --proxy.proxyenodeurlpair is elided and
there's an expectation that geth will fail to start.