txpool: contract call caching & remove gpm check (#1499)
**Contract Call Caching**

This PR adds a caching layer to the txpool. It works as follows:

Cache has 2 parts:

BlockContext which represents contract's information required to process a block. (gasPriceMinimum per currency, whitelistedCurrencies, gasForAlternativeCurrency)
CurrencyManager object which is able to make currency conversion, and maintains al local cache of exchange rates
Both objects are part of the txpoolContext, and they are recreated with every reset(). This means they are recreated everytime the txpools learns about a new block.

**TXPool Behaviour Changes**

1. txpool no longer checks for gas price minimum (see #1499 (comment))
2. If getCurrencyWhitelist() fails, then assume no custom fee currency
3. is whitelisted (before: any fee currency was whitelisted)