Stop replicas from sending enode certs to proxies (#1507)
### Description

Stop replicas from sending enode certs to proxies. `proxiedValidatorEngine.sendValEnodesShareMsg`
did check if the node was the primary, but `sendEnodeCerts` did not. This matches which messages
are sent as a replica in `announce.go`.

This will stop the error `invalid enode certificate message map because of old version`
from occurring on proxies with replicas.

### Tested

No

### Related issues

- Partial Fix of  #1350

### Backwards compatibility

Yes