Decode inner message on Message.FromPayload
Message decoding is split between many places in the code,
this leads to needing to decode messges while deep down the call stack
meaning the potential errors from failing to decode need to be handled
at each level of the call stack, thus complicating code. This change
means that we decode in one place only, which allows us to limit the
places that need to handle decode errors.