Fix default DatabaseCache value.
In the PR #18087 (https://github.com/ethereum/go-ethereum/pull/18087/files),
the trie cached was splited between Clean and Dirty.
As the cache was suppose to be 1gb, and the Dirty cache was assigned
using part of the DatabaseCache, it means that the DatabaseCache should
be updated too, to validate that all the sum of the caches are still
1gb.
This was missed in an early merge, and fixed here.