Removes dependency from vm to consensus
1) Split consensus ChainReader into ChainHeaderReader (done in upstream)
and ChainContext (which inherits from ChainReader and adds EVMRunner
factory methods)

2) Modifies all consensus.Engine methods either to remove the
ChainReader parameter or change it to ChainHeaderReader. Potentially the
parameter can be removed since Istanbul engine already receives the
ChainContext on creation (with SetChain())

3) Moves vm.ChainContext to core.ChainContext. This is the original
location in upstream. And helps break the dependency

4) Duplicates core.ChainContext in vmcontext.ChainContext to break
dependency. core MUST depende on vmcontext, so we have to avoid the
inverse. For that, instead of expecting a core.ChainContext argument, it
now expects its own locat interface which is a copy of the former