Refactors contract call layer (#1513)
Completely rewrites the low level contract API. This is done in a backwards compatible manner, since the current contract api is maintained, and rewritten so that it uses the new api. Thus, all users of the contract API are not required to migrate immediately, and we can do that migration in steps.

#### Low Level Contract API

Rewrite the low level contract API, by defining the following types

1) `EVMRunner` interface to do write/read calls to the state using an EVM


```go
// EVMRunner provides a simplified API to run EVM calls
// EVM's sender, gasPrice, txFeeRecipient and state are set by the runner on each call
// This object can be re-used many times in contrast to the EVM's single use behaviour.
type EVMRunner interface {
	// Execute performs a potentially write operation over the runner's state
	// It can be seen as a message (input,value) from sender to recipient that returns `ret`
	Execute(recipient common.Address, input []byte, gas uint64, value *big.Int) (ret []byte, leftOverGas uint64, err error)

	// Query performs a read operation over the runner's state
	// It can be seen as a message (input,value) from sender to recipient that returns `ret`
	Query(recipient common.Address, input []byte, gas uint64) (ret []byte, leftOverGas uint64, err error)

	// StopGasMetering backward compatibility method to stop gas metering
	// Deprecated. DO NOT USE
	StopGasMetering()

	// StartGasMetering backward compatibility method to start gas metering
	// Deprecated. DO NOT USE
	StartGasMetering()

	// GetStateDB retrieves current stateDB within the runner
	GetStateDB() StateDB
}
```

 * Provides a way to make multiple calls using the same object. the EVM object should not be reused!
 * It's an entry point to implement low level cache to EVM calls, since it requires only implementing a Decorator object that first checks the cache, and on miss call the decoratee.
 * `SharedEVMRunner` implementation is provided since we DO reuse the evm, and changing that would be a HF, so we need to that later
 * Exposes gas metering methods also as a backward compatible measure that we need to remove later

2) To create `EVMRunner` we expose:
```go
func NewEVMRunner(chain ExtendedChainContext, header *types.Header, state vm.StateDB) vm.EVMRunner

func GetSystemEVMRunnerFactory(chain ExtendedChainContext) func(header *types.Header, state vm.StateDB) (vm.EVMRunner, error)
```

The second one, is for backward compatibility with contract_comm and will be removed on a later PR

On https://github.com/celo-org/celo-blockchain/pull/1522 we introduce utility factories on `Blockchain` and `Lightchain` who already provide the `ExtendedChainContext` and thus only the header and state are needed.

3) On `contracts` module we expose `Method` and `BoundMethod`. The first represents a method can be called once it's binded to an address with `Bind(common.address)`.

The `BoundMethod` exposes:
* `Query(vmRunner vm.EVMRunner, result interface{}, args ...interface{}) error`: this runs the method a a query (read-only) over the given runner and parameters
* `Execute(vmRunner vm.EVMRunner, result interface{}, value *big.Int, args ...interface{}) error`: this methods runs a write operation over the given runner and parameters

Nice property about this object is that can be statically created and reused. So work as configuration for method call.

#### ChainContext and ExtendedChainContext

In the past we've used & abused `ChainContext` as an object that is available, and so we added methods to it for different objects to access it.

I would like to revert that decision, and have minimal changes to the ChainContext. I've created `ExtendedChainContext` (inherits from ChainContext) as the interface needed to create an `EVMRunner`. So that methods that are only needed for that object belong there, and not to a commonly used interface as `ChainContext`

#### Changes on Error logging

On master currently logging on error happens 3 times.

The typical call stack is:
  1. `makeCallWithContractId()`
  2. `makeCallFromSytem()`
  3. StaticCallFromSystem or CallFromSystem (that both call `handleAbiCall()`

On error we end up logging 3 TIMES!

  1. First [contracts/utils.go:handleAbiCall](https://github.com/celo-org/celo-blockchain/blob/master/contracts/utils.go#L49-L88) will log using `trance` or `error` level depending on `methodName` or type of error.
  2. On [makeCallFromSystem()](https://github.com/celo-org/celo-blockchain/blob/master/contract_comm/evm.go#L117-L120) we log with error level
  3. On [makeWallWithContractId()](https://github.com/celo-org/celo-blockchain/blob/master/contract_comm/evm.go#L117-L120) we log again with error level
  
This PR changes that. We know **only log errors** on `Call.Run()`. And it's always on `error` level.