Removes the miner's pending cache (#1555)
This modifies the miner <-> engine interface to remove the results channel from the engine to the miner.
With #1394, validators save the results of processing the block
and use them while inserting the block. This cache is the the engine. Prior to this change, there was a similar
cache in the miner, but it would only be hit when the block was proposed by the node (1/100 hit rate). This PR
removes this cache and the communication channel back from the engine to the miner. The miner submits
requests to the engine, but no longer receives any information back from the engine once the block is committed.

Thus this simplifies the miner and the interface between the engine and the miner.