Remove support for celo/64 and celo/65 protocols. (#1565)
Now that Donut has been activated on Alfajores, Baklava, and mainnet, all nodes which are still able to connect and sync are running v1.3.x and thus support celo/66.  Therefore, it is possible to remove support for celo/64 and celo/65 and with them the notion of a "primary" protocol.  See #1213.