Migrates all contractst calls to EVMRunner (#1522)
This PR removes contract_comm modules, and recreates them into contracts. On that move, all calls are now done using contract.BoundMethod and they expect an vm.EVMRunner instance instead of a (header,state) pair.

Doing so, implies that contract callers now require the ability to create an vm.EVMRunner which in the past was done thanks to a singleton living in contract_comm/evm.go. On this PR, that responsibility is on Blockchain and Lightchain object. These 2 objects are easily available in most of the other packages in some form of another, so migration is direct is most of the cases.