Move several sign&address properties from istanbul.Backend to its own type (#1597)
Properties received in the `Authorize(...)` method of `istanbul.Backend` are guarded by a lock to avoid race conditions during  the setting of the values (golang reference assigmnent is not guaranteed to be atomic I believe).

I moved all of those into a type `AuthorizeInfo` and used an `atomic.Value` to be able to remove the lock. The `decryptFn` was not being guarded by the lock, as many other `address` fields, in what I believe it was a possible race condition bug during assignment and check. It's probably the case that it would have never been triggered in real world use though.

I also incremented a sleep in an unrelated test since the wait period for a geth instance to start was too low and it was making the test a bit flaky.