Remove 'Header' field from vm.Context.
This was added in PR #317 but was never set in `vm.NewEVMContext()`.  As a result, it was always nil.  After the refactoring of contract calls to use the vm.EVMRunner types, we now don't use it outside of eth/api_tracer.go.  And the use there is now broken because the field is nil: when tracing a transaction or block the `vmRunner` created for the contract calls will use the header and state from the current chain head rather than from the specified block.  This fixes that by removing the field and updating the tracing code to not use it either.