Announce protocol code refactor (#1590)
Extracts most of the Announce protocol behaviour into the `AnnounceManager` type. This is a partial PR, in the sense that most work is yet to be done to completely move all announce code into the announce manager, but since it's already a really big change, I decided to merge this and continue working on it on a different branch.

This PR extracts a big part of the announce code into the new announce manager type. Work yet to be done is:
- finish removing everyting from backend into announcemanager
- decide on what the public interface of the announcemanager will be, once everything is in place
- minimize the dependencies on the announce manager constructor
- improve the unit testing on it, using the advantage of it being decoupled from istanbul.Backend

Some previous PRs were written in orden to make this easier (GossipCache, the atomic.Value for wallets).