Go back to using the transfer's sender in Tobin tax calculation. (#1611)
Even though the Tobin tax calculation doesn't depend on the sender, the change (in PR #1465) of the sender from the transaction's sender to the zero account led to a bad block error (i.e. a hard fork change).  This was due to a conjunction of factors including the upstream implementation of EIP-161 not being quite correct for the case of Celo (because of the existence of system EVM calls made from celo-blockchain).  This commit reverts the change, by adding an `ExecuteFrom()` method to the EVMRunner type.