Update prepare-system-contracts make rule
This commit makes two changes:

1. Instead of symlinking to the celo-monorepo root from the
celo-blockchain root, we now copy just to the compiled contracts from
the monorepo to compiled-system-contracts. This should help our tooling
by reducing the amount of files in scope of searches etc. Previously
certain tooling would follow the symlink and end up processing all files
in celo-monorepo resulting in significant slowdown.

2. The overrides for the make rule (MONOREPO_PATH & MONOREPO_COMMIT) now
are not persistent and need to be provided on all invocations of the
make rule. This change was made because the previous implementation
could end up functioning in strange ways due to the statefulness. This
approach is just more straightforward.

CI has also been updated to handle the new path for the contracts.