Update prepare-system-contracts make rule
This commit makes two changes:

1. Instead of symlinking to the celo-monorepo root from the
celo-blockchain root, we now symlink just to the compiled contracts.
This should help our tooling by reducing the amount of files in scope of
searches etc.

2. The overrides for the make rule (MONOREPO_PATH & MONOREPO_COMMIT) now
are not persistent and need to be provided on all invocations of the
make rule. This change was made because the previous implementation
could end up functioning in strange ways due to the statefulness. This
approach is just more straightforward.