Consolidate istanbul.Message encoding/decoding (#1592)
Message decoding and encoding, prevously happend in 2 steps. This commit
makes decoding and encoding a message an atomic operation, either it
succeeds or it fails.

This allows us to decode in just one place in the code, allowing us to
have a consistent approach to decode errors. By being able to respond to
decode errors at a higer level in the stack, we remove the need to pass
errors back up the stack to indicate invalid messages. It also helps
separate encoding and decoding logic from our domain logic.