Fix VersionCertificate rlp encoding/decoding (#1622)
PR #1592 introduced a backwards incompatibilty because it changed the
rlp structure that version certificates were stored and retrieved with
in the version certificate db.

Prior to #1592, there were two types: VersionCertificateEntry (used for
encoding/decoding to RLP for the database) and VersionCertificate (used for
communicating with peers), and each had their own RLP encode/decode
implementation Consolidate istanbul.Message encoding/decoding. #1592
consolidated them into a single type called VersionCertificate but broke
compatibility with the encoding that was used for the database.

This PR keeps a single type but reinstates the previous rlp structure in the
version cert db by introducing the functions encodeVersionCertificate and
decodeVersionCertificate, which enforce the old serialisation format.