Add support for "system calls receipt" in eth_getTransactionReceipt
If systems calls made by celo-blockchain result in logs, we add a receipt including those logs, which lists the block's hash as the transaction hash field on the receipt.  This PR adds support to eth_getTransactionReceipt to return this receipt when given a block hash.  One key difference, however, is that no receipt is added in case there were no logs, whereas eth_getTransactionReceipt will return an empty receipt in such cases, so that the case of no logs can't be confused with the case of not having the block.  We may in the future wish to change it so that the receipt is also created during block processing, but that would be a hard fork change.