core/types: Reconcile transaction formats
This has some minor deletions from upstream to keep the block format the same
and then major changes around the transaction format.

Upstream the transaction there is now a transaction interface with three
implementations:
- Legacy Tx      (0x00)
- Access List Tx (0x01)
- Dynamic Fee Tx (0x02)

I added the following transaction type
- Legacy Celo Tx

This is designed such that we don't need to overload Legacy Tx with
the eth compatible field (related to it's RLP encoding).

Note: This design may have to be reworked to handle legacy transactions
more cleanly.