Move the statedb prepare call for the block receipt to inside Finalize().
This call needs to be done during both block construction and block verification. It was previously only done during verification, but this didn't cause problems because of a statedb .Copy() during consutrction, in between transaction processing and the Finalize().  This commit moves it to the right place, ensuring it's done in both scenarios without relying on unrelated code.