Remove unnecessary statedb copies and move the prepare inside Finalise (#1673)
* Move the statedb prepare call for the block receipt to inside Finalize().

This call needs to be done during both block construction and block verification. It was previously only done during verification, but this didn't cause problems because of a statedb .Copy() during consutrction, in between transaction processing and the Finalize().  This commit moves it to the right place, ensuring it's done in both scenarios without relying on unrelated code.

* Remove superfluous statedb Copy() calls.

The one in block.go is superfluous now that the `Prepare()` has been moved into `Finalize()`.  The one in backend.go was always superfluous, as the statedb instance there is a fresh one obtained a few lines above.

* Factor out "block receipt" into a function.

This reduces duplication and ensures the same logic is used in all cases.