Fix deadlock during StartValidating
StartValidating makes a call to RefreshValPeers while holding
coreMu and RefreshValPeers waits for all validator peers to be deleted
and then reconnects to known validators.

If any of those peers has called IsValidating before RefreshValPeers
tries to delete them, the system gets stuck in a deadlock because
IsValidating also tries to acquire coreMu. The peer will never acquire
coreMu because it is held by StartValidating, and StartValidating will
never return because it is waiting for all peers to disconnect.

This commit makes coreStarted into an atomic variable so that peers can
make threadsafe calls to IsValidating without needing to acquire
coreStarted.