Fix long wait for nodes to connect
At test startup sometimes nodes were taking in the region of 30s to connect
whilst other times it was happening in μs. The problem was we were
trying to connect all peers to all other peers. That meant that for any
two peers they would both dial each other. Sometimes if this occurred
close enough in time both sides would hang up the connections (I call
this cross dialling). This happens because each side counts their
outgoning connection as connected and then when the incomming connection
arrives they drop it because they see themselves as already connected.
When this happened nodes would retry after some time probably 30s and
then be connected.

The fix was to ensure that for any two nodes only one of them dials the
other.