Adjust the way config is handled in test package
Because of the duplication of istanbul config in the codebase,
(see https://github.com/celo-org/celo-blockchain/issues/1693)
when using the test package it was difficult to understand which config
should be used and possible to end up with conflicting intanbul config
in different parts of the system.

This commit makes it clear where the config should be set and enforces
consistent config across the two config objects.