Wip debugging
So now I have a new test that shows that the announce protocol, causes
connections to be forcibly dropped and remade, but this causes issues
because of the dialHistoryExpiration and inboundTrhottleTime being 35s
and 30s respectivley. The announce message sent at 5 second intervals
causes the connections to be hung up and then the connections don't get
established again for 35 or 30 seconds.