Fix race condition on core blockchain wg
When using a wg Add is intended to be called before wait, calling Add
and Wait concurrently leads to data races, and the blockchain calls Add
all over the place in many public methods that can be called from other
threads.

The solution is to lock over calls to Add and Wait.