Revert "Remove unnecessary statedb copies and move the prepare inside Finalise (#1673)"
This reverts commit 8b8e717689daaac8a854aa931df07009e7be0a34.
The combination of v1.10.7 and the reverted commit caused an issue in the
e2e governance test. The test failed when it was looking for a specific event
that is produced during the "block transaction" performed at the end of the
block.

I'm not sure whey the update to v1.10.7 caused this change to no longer
work.