Move extra check message processing to handleMsg
Move the extra processing on the result of checkMessage from the
preprepare handler up to handleMsg.

Remove the call to checkMessage from the preprepare handler, this
checking is now performed in handleMsg.

Remove the test for oldMessgage on the preprepare handler.

Also combine the individual checks for old sequence and old round into
a single view comparison.