Move extra checkMessage processing to handleMsg
Move the extra processing that the commit message handler was doing in
response to the result from checkMessage into handleMsg.

Also remove the call to check message from the commit handler, since
all those checks are now done in handleMsg.

And also removed the test cases for future and old messages and commits
for the previous sequence from the commit handler test, since that
functionality is now covered by handleMsg.