miner: Properly stop the prefetcher
The makes it such that the prefetcher associated with a stateDB object
is always stopped once the state associated with it is done.

Note: The prefetcher could probably be ran for a longer period of time, but
there's no other clean location to stop it at while properly managing the
lifetime of the object.