Move extra prepare validation into handleMsg
Moved the validate logic for prepares messages into handleMsg.

Since in handleMsg at the point we apply the extra validation logic we
have already checked that the prepare message's view matches that of the
core instance we only now need to check that the digest matches the core
instance's current proposal. This meant we could do away with the
verifyPrepare method and just check the digest.

Moved the test case for mistmatching digest into handler_test.go from
prepare_test.go.

Also was now able to remove the TestVerifyPrepare test since the method
no longer existed.