Move extra commit validation into handleMsg
Moved the validate logic for commit messages into handleMsg.

Since in handleMsg at the point we apply the extra validation logic we
have already checked that the commit message's view matches that of the
core instance we only now need to check that the digest matches the core
instance's current proposal. This meant we could do away with the
verifyCommit method and just check the digest.

Moved the test case for mistmatching digest into handler_test.go from
commit_test.go.

Also was now able to remove the TestVerifyCommit test since the method
no longer existed.

Also added a couple of new test cases to check for invalid committed and
epoc seals.