Avoid calling NextBlockValidators for preparedCert
NextBlockValidators does processing to update the validator set in the
event that validators have been added or removed. In doing this it makes
contract calls.

Given that prepared certs are only used in the context of round changes,
the current validator set will never be different from the set used in
round 0 so the call to NextBlockValidators can be avoided.