Update prepare test to go through handleMsg
Rather than calling handlePreprepare the test now calls handleMsg
meaning the messages experience the full set of processing logic.

As part of this change I removed the "test existing block" test case
since it does not make sense any more. This is because preprepares for
an existing block, by definition are old, and the validation in handleMsg
will return an error. This error case is already covered by the
TestHandleMsg test.

As part of removing the "test existing block" case I noticed that there
was an error in the testing logic such that all checks below the
'continue' statement was never executed, these checks have now been
re-instated.

Also updated the preprepare benchmark to go through handleMsg.