Consolidate unit test CI jobs
We had a job to run unit-tests and then essentially the same job to run
them with coverage. The running time for these jobs was about equal,
sometimes one would be a bit faster and vice versa.

This PR removes one of the jobs leaving us with just one unit-test job
that tracks coverage.

This has the benefit of halving the chance that a build gets failed due
to a flaky test.

Also standardised the approach to running tests between the remaining
unit test jobs.