Fix JSON RPC return + gas estimation
The first change is passing parameters through to gas estimation. This is
important, specifically that if the fee currency parameter is not provided,
gas estimation will fail when it should not if the EOA does not have any
native currency, but does have a whitelisted fee currency and is trying to pay
for fees in that.

The second change includes the CeloDynamicFeeType transactions when
constructing responses over the JSON RPC.