Improve duplicate block insert logging (#1780)
The previous logging would output a warning for any case where we tried
to insert a block with a mismatching parent hash. The problem with this
was that there were cases where that is expected and also cases where
that was clearly an error and it wasn't distinguishing between them.

We now return errors in the error cases and log an info when an expected
duplicate block insert occurs.