Revise use of Msgs, replace Events with PendingEvent
Defining Msgs as a variable that contined messages for the current
height was unnecessary and introduced confusion over how it was set.

Now Msgs simply holds all the messages sent and received. Because of
this removed the statement that there is no shared state between
heights, which was already not correct because the current height was
state that was shared between heights.

Ensure that Msgs is used in quorumRound and f1Round.

Events was a set of events that at most contained one event so changed
it to be a variable rather than a set.