Eth api using baseFee only when needed (#1818)
The Problem
The eth RPC API after expreso required some amending.

On ethereum, baseFee ( called gasPriceMinimun on Celo) is present on the block header. For Celo, it's present on the GasPriceMinimum contract, and it also needs to be denominated in all different fee currencies.

This difference poses a problem on the RPC, since whenever a Transaction or a TransactionReceipt is constructed, the gasPrice field is computed for new transaction. To do so, it uses the baseFee. Obtaining the baseFee for block X is easy on Ethereum and hard on Celo; since it's a smart contract call, the node needs to have the state available but that's not true for historical blocks on any node that it's not an archive node.

The Solution
To solve this issue, this PR takes the following approach:

If node has access to the block's state, then gasPrice is computed as usual
If node doesn't have access to block's state, then gasPrice is nil
Additionally, we have the same problem with the new AccessList rpc endpoint, that computes the access list for a transaction on a given block. If the block is too old, and we don't have the state, then the rpc will fail.