Cherry pick #1840 to master (#1842)
### Description

_This PR cherry picks #1840 to `master`. A copy of the description is included here_

In the course of getting Baklava into condition to restart block production, I noticed that the
miner would fail to initialize on nodes that had replaced there chain data from backup. The observed
error was that the randomness commitment cache would fail to be populated, which is caused by
replacing the chain data on that node. This error is supposed to be recoverable, but no code
existed on the code path that was failing to do so. Recovery would only occur after syncing to the
network, which these nodes did not do.

This PR adds a conditional to try recovering the randomness cache once if there is a cache miss
during miner initialization.

### Other changes

* Changed the version number of 1.6.0-unstable to align with our stated release process.

### Tested

* Changes tested on Baklava